requirements.txt
discord.py
databases[postgresql]
python-dotenv
psutil
requests
python-dateutil